Changed id from int to long for scalability and added empty check for input parameters

+++ src/main/java/com/bitespeed/contact_consolidation/dto/ContactResponseDTO.java
@@ -9,11 +9,11 @@ import java.util.List;
 @AllArgsConstructor
 public class ContactResponseDTO {
 
-    private int primaryContactId;
+    private long primaryContactId;
 
     private List<String> emails;
 
     private List<String> phoneNumbers;
 
-    private List<Integer> secondaryContactIds;
+    private List<Long> secondaryContactIds;
 }
